composer: remove and ignore composer.lock
It's possible to use it for libraries, but not necessary and I believe
it's better to not have it.

Libraries should have the broadest support of dependencies so actual
users of the libraries face less version constraints. This should also
apply to developers of this library.

The composer.lock _only_ applies when developing this library and it's
irrelevant for users of this library.