Remove file path translation
The `realpath()` currently causes issues when testing, as in tests
we mock the `copy` and don't perform it, leading `realpath()` being
called on a non-existent path and thus passing `false` to `str_replace`
as 3rd arg and thus breaking it in PHP 8.

This will impact the functionality, just the output won't be as nice;
at least for now.