Return messages directly and not the bag
I don't think the intended output was ever to have the MessageBag
being an intermediary for serializing the response, when directly
returning the messages yields the same result.

The gracious changes in the tests result from not having the MessageBag
available anymore and making the shortcut just comparing the whole
response.

IMHO it also gives a clearer indication how such a thing looks like
(the shape) and find the individual probes into the result harder to
grasp in the big picture.