tests: switch to json requests to improve DX
This has the benefit in case of error/exceptions during runtime, we
don't get back the HTML error page but a JSON serializable
representation aka more readable for devs.

I'm aware that technically the $response is a `TestResponse`, but for
all intents and purposes, in this context we just use it as JsonReponse.